Backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
distro==1.9.0
exceptiongroup==1.3.0
fastapi==0.109.2
h11==0.16.0
httpcore==1.0.9
httpx==0.26.0
idna==3.10
iniconfig==2.1.0
openai==1.12.0
packaging==25.0
pluggy==1.6.0
pydantic==2.6.1
pydantic-settings==2.1.0
pydantic_core==2.16.2
PyMuPDF==1.23.8
PyMuPDFb==1.23.7
pyngrok==7.2.8
pytest==8.0.0
pytest-asyncio==0.23.5
python-dotenv==1.1.0
python-multipart==0.0.9
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
sniffio==1.3.1
starlette==0.36.3
tiktoken==0.6.0
tomli==2.2.1
tqdm==4.67.1
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.27.1
